Update userguide with new expected JSON payloads
Additionally, add links to related sections in swagger such that users
can search the full documentation. We typically have the examples in
place to walk the user through a contrived process to using piazza, so
having links to the full API documentation is necessary to show possible
and expected fields, other uses of methods, etc.